1.21.8 (#1)
* Add support for 768 and 769 protocols //Outdated

* Add support for 768 and 769 protocols //Outdated

* improve the packet handling, now packets handled //Outdated

* Refactor other proto: 768, 767, 766, 765, as done for 769 //Outdated

* New profiling options and performance improvements for world handling

* Fast cookie handling and fix networking issues
    - now plugins can subscribe to specified cookie and read from it
    - fixed sending redundant info through network when sending registers
    - now tags sent to clients
    - improve datapacks handling

* Parallelize and optimize ticking

* Implement 1.21.5 support and removed old protocols
  - It's hard to support conversation for resource data for multiply protocols, so i decided to support only latest release
  - IDK should i keep the tools for supporting multiply protocols for plugin developers?
  - Use modified extractor from PumpkinMC to get more data for entities and blocks

* Refactor storage::world_data and add tags api

* Bug fixes, new features
Features:
  - new tags api
  - add metadata to entities
  - added compression to resources
  - added enbt support for resources
  - Effects is now partially handled for entities
Fixes:
  - fixed virtual client packets parsing
  - Open callbacks in world_data
Tweaks:
  - weaken requirement for chunk encoding, now allowed dynamic array instead of fixed_array
  - Now default light processor sets maximal light level to every block instead of doing noting
  - replaced int8_t to base_objects::packets::effect_flags in entityEffect packet
  - Implemented enable_entity_light_source_updates and enable_entity_light_source_updates_include_rot for light processors
  - Removed enable_entity_updates_when_hold_light_source for light processors

* Refactor networking and bug fixes
  - Moved networking from boost::asio to fast_task::networking
  - Now tcp server implemented as plugin
  - moved protocolHelper/packets.hpp to api/packets.hpp
  - moved protocolHelper files to built_in_plugin and refactored them
  - now virtual clients has special flag and are not sent to clients
  - refactored SpecialPluginHandshake, now implemented using plugin
  - all sync_event, event and callback now inherit base_event class, this allows for plugins use them same as event
  - new increase_time flag in world metadata
  - new server api for shutdown
  - removed asio api

* Refactor and bugfixes
  - allow players enter loading screen
  - optimize resources loading, 5s debug and 0.9s Release
  - preparing interface for player entity to pass data to clients
  - fix nbt encoding
  - preparations for incremental world generation

* Improve networking and other features
  * add log file
  * add more options in configuration
  * add send_indirect to api::network::tcp::session
  * improve keep alive handling

* Bug fixes, removed multiple protocol support and improve packet handling

* Plugins cleanup

* Add custom config options for config.json and use this in allowlist plugin
 - also use client_data_holder instead of string in api::allowlist::on_kick

* Bump a big refactor, fixed a pile of bugs
 - also added feature to Chat struct, look at from_enbt_with_format
 - used new functions in OpenSSL instead of deprecated ones

* Replace fast_task/src/files and fast_task/src/networking to fast_task/include

* Use new data from extractor

* Prepare to new_packets api to replace old one
new api uses reflection to encode and decode packets
- the virtual_client now uses one callback function to process all packets
- added move variant for sendPacket in SharedClientData
- also update libraries

* Prepare nearly all packets and implement serialization of packet to string

* prepare for 1.25.8
 - update plugins and use `new_packets` api
 - use new plugin naming
 - removed `protocol` api for `new_packets` api
 - removed `packets` api for new one
 - fixed some packets

* Complete 1.21.8 protocol implementation
 - changed reflection library to own tooling, new reflect_map tool to parse headers of structs and provide header for iterating structs and serializing enums(a lot faster and less bugs)
 - change new_packets api name to packets
 - partial implementation for selector
 - added std::hash for id_source
 - added api::permissions::is_in_group function
 - moved all api cpp files to api/bin
 - SharedClientData now directly stores compatible_plugins instead of names
 - changed std::vector in api::plugin to list_array
 - added new bounds: cubic_bounds_block_radius_out, spherical_bounds_chunk_out and spherical_bounds_block_out

* Add license

* Refactor used clang to fix some bugs
 but failed to link because vcpkg not using clang-cl
 also fixed stack unwind for windows while using log::*
   now instead of running functions in place, the message queued for execution

+++ src/base_objects/number_provider.cpp
@@ -1,4 +1,11 @@
-
+/*
+ * Copyright 2024-Present Danyil Melnytskyi. All Rights Reserved.
+ *
+ * Licensed under the Apache License 2.0 (the "License"). You may not use
+ * this file except in compliance with the License. You can obtain a copy
+ * in the file LICENSE in the source distribution or at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ */
 #include <random>
 #include <src/base_objects/number_provider.hpp>
 
@@ -54,10 +61,10 @@ namespace copper_server::base_objects {
                     return std::uniform_real_distribution<float>(arg->get_min_inclusive_float(), arg->get_max_exclusive_float())(engine);
                 } else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_clamped_normal*>) {
                     auto engine = std::default_random_engine();
-                    return std::clamp<float>(std::normal_distribution<float>(arg->mean, arg->deviation)(engine), arg->min, arg->max);
+                    return std::clamp<float>(std::normal_distribution<float>(arg->mean, arg->deviation)(engine), (float)arg->min, (float)arg->max);
                 } else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_trapezoid*>) {
                     auto engine = std::default_random_engine();
-                    return std::clamp<float>(std::uniform_real_distribution<float>(arg->min, arg->max)(engine), arg->min, arg->max);
+                    return std::clamp<float>(std::uniform_real_distribution<float>((float)arg->min, (float)arg->max)(engine), (float)arg->min, (float)arg->max);
                 } else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_clamped*>)
                     return std::clamp<float>(arg->source.get_float(), arg->get_min_inclusive_float(), arg->get_max_inclusive_float());
                 else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_weighted_list*>) {
@@ -77,7 +84,7 @@ namespace copper_server::base_objects {
                     return std::clamp<float>(std::uniform_real_distribution<float>(arg->get_min_inclusive_float(), arg->get_max_exclusive_float())(engine), arg->get_min_inclusive_float(), arg->get_max_exclusive_float());
                 } else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_binomial*>) {
                     auto engine = std::default_random_engine();
-                    return std::binomial_distribution<int32_t>(arg->n.get_int(), arg->p.get_int())(engine);
+                    return (float)std::binomial_distribution<int32_t>(arg->n.get_int(), arg->p.get_int())(engine);
                 }
                 //else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_score*>)
                 //    return arg->source.get_float();
@@ -102,7 +109,7 @@ namespace copper_server::base_objects {
                     return std::uniform_int_distribution<int32_t>(arg->get_min_inclusive_int(), arg->get_max_exclusive_int())(engine);
                 } else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_clamped_normal*>) {
                     auto engine = std::default_random_engine();
-                    return std::clamp<int32_t>(std::normal_distribution<float>(arg->mean, arg->deviation)(engine), arg->min, arg->max);
+                    return std::clamp<int32_t>((int32_t)std::normal_distribution<float>(arg->mean, arg->deviation)(engine), arg->min, arg->max);
                 } else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_trapezoid*>) {
                     auto engine = std::default_random_engine();
                     return std::clamp<int32_t>(std::uniform_int_distribution<int32_t>(arg->min, arg->max)(engine), arg->min, arg->max);
@@ -150,10 +157,10 @@ namespace copper_server::base_objects {
                     return std::uniform_real_distribution<float>(arg->get_min_inclusive_float(), arg->get_max_exclusive_float())(engine);
                 } else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_clamped_normal*>) {
                     auto engine = std::default_random_engine();
-                    return std::clamp<float>(std::normal_distribution<float>(arg->mean, arg->deviation)(engine), arg->min, arg->max);
+                    return std::clamp<float>(std::normal_distribution<float>(arg->mean, arg->deviation)(engine), (float)arg->min, (float)arg->max);
                 } else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_trapezoid*>) {
                     auto engine = std::default_random_engine();
-                    return std::clamp<float>(std::uniform_real_distribution<float>(arg->min, arg->max)(engine), arg->min, arg->max);
+                    return std::clamp<float>(std::uniform_real_distribution<float>((float)arg->min, (float)arg->max)(engine), (float)arg->min, (float)arg->max);
                 } else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_clamped*>)
                     return std::clamp<float>(arg->source.get_float(other_data), arg->get_min_inclusive_float(), arg->get_max_inclusive_float());
                 else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_weighted_list*>) {
@@ -173,7 +180,7 @@ namespace copper_server::base_objects {
                     return std::clamp<float>(std::uniform_real_distribution<float>(arg->get_min_inclusive_float(), arg->get_max_exclusive_float())(engine), arg->get_min_inclusive_float(), arg->get_max_exclusive_float());
                 } else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_binomial*>) {
                     auto engine = std::default_random_engine();
-                    return std::binomial_distribution<int32_t>(arg->n.get_int(other_data), arg->p.get_int(other_data))(engine);
+                    return (float)std::binomial_distribution<int32_t>(arg->n.get_int(other_data), arg->p.get_int(other_data))(engine);
                 }
                 //else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_score*>)
                 //    return arg->source.get_float(other_data);
@@ -198,7 +205,7 @@ namespace copper_server::base_objects {
                     return std::uniform_int_distribution<int32_t>(arg->get_min_inclusive_int(), arg->get_max_exclusive_int())(engine);
                 } else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_clamped_normal*>) {
                     auto engine = std::default_random_engine();
-                    return std::clamp<int32_t>(std::normal_distribution<float>(arg->mean, arg->deviation)(engine), arg->min, arg->max);
+                    return std::clamp<int32_t>((int32_t)std::normal_distribution<float>(arg->mean, arg->deviation)(engine), arg->min, arg->max);
                 } else if constexpr (std::is_same_v<std::decay_t<decltype(arg)>, number_provider_trapezoid*>) {
                     auto engine = std::default_random_engine();
                     return std::clamp<int32_t>(std::uniform_int_distribution<int32_t>(arg->min, arg->max)(engine), arg->min, arg->max);
